+++ media/lua/client/CHC_config.lua
@@ -34,7 +34,32 @@ CHC_settings = {
     }
 }
 
+local init_cfg = {
+    recipe_selector_modifier = 1, -- none
+    category_selector_modifier = 1,
+    tab_selector_modifier = 1,
+    list_font_size = 3, -- large
+    show_icons = false,
+    allow_special_search = true,
+    show_hidden = true,
+    close_all_on_exit = false,
+    main_window = { x = 100, y = 100, w = 1000, h = 600 },
+    uses = { sep_x = 500, filter_asc = true, filter_type = "all" },
+    craft = { sep_x = 500, filter_asc = true, filter_type = "all" },
+    search = {
+        items = { sep_x = 500, filter_asc = true, filter_type = "all" },
+        recipes = { sep_x = 500, filter_asc = true, filter_type = "all" }
+    },
+    favorites = {
+        items = { sep_x = 500, filter_asc = true, filter_type = "all" },
+        recipes = { sep_x = 500, filter_asc = true, filter_type = "all" }
+    }
+}
+
 local function onModOptionsApply(values)
+    CHC_settings.config.recipe_selector_modifier = values.settings.options.recipe_selector_modifier
+    CHC_settings.config.category_selector_modifier = values.settings.options.category_selector_modifier
+    CHC_settings.config.tab_selector_modifier = values.settings.options.tab_selector_modifier
     CHC_settings.config.list_font_size = values.settings.options.list_font_size
     CHC_settings.config.allow_special_search = values.settings.options.allow_special_search
     CHC_settings.config.show_icons = values.settings.options.show_icons
@@ -84,6 +109,36 @@ if ModOptions and ModOptions.getInstance then
                 default = 3,
                 OnApplyMainMenu = onModOptionsApply,
                 OnApplyInGame = onModOptionsApply
+            },
+            recipe_selector_modifier = {
+                getText("IGUI_None"), getText("UI_optionscreen_CycleContainerKey1"),
+                getText("UI_optionscreen_CycleContainerKey2"),
+                getText("UI_optionscreen_CycleContainerKey3"),
+                name = "IGUI_RecipeSelectorModifier",
+                tooltip = getText("IGUI_RecipeSelectorModifierTooltip", getText("UI_optionscreen_binding_chc_move_up"), getText("UI_optionscreen_binding_chc_move_down")),
+                default = 1,
+                OnApplyMainMenu = onModOptionsApply,
+                OnApplyInGame = onModOptionsApply
+            },
+            category_selector_modifier = {
+                getText("IGUI_None"), getText("UI_optionscreen_CycleContainerKey1"),
+                getText("UI_optionscreen_CycleContainerKey2"),
+                getText("UI_optionscreen_CycleContainerKey3"),
+                name = "IGUI_CategorySelectorModifier",
+                tooltip = getText("IGUI_CategorySelectorModifierTooltip", getText("UI_optionscreen_binding_chc_move_left"), getText("UI_optionscreen_binding_chc_move_right")),
+                default = 1,
+                OnApplyMainMenu = onModOptionsApply,
+                OnApplyInGame = onModOptionsApply
+            },
+            tab_selector_modifier = {
+                getText("IGUI_None"), getText("UI_optionscreen_CycleContainerKey1"),
+                getText("UI_optionscreen_CycleContainerKey2"),
+                getText("UI_optionscreen_CycleContainerKey3"),
+                name = "IGUI_TabSelectorModifier",
+                tooltip = getText("IGUI_TabSelectorModifierTooltip", getText("UI_optionscreen_binding_chc_move_tab_left"), getText("UI_optionscreen_binding_chc_move_tab_right")),
+                default = 1,
+                OnApplyMainMenu = onModOptionsApply,
+                OnApplyInGame = onModOptionsApply
             }
         },
         mod_id = "CraftHelperContinued",
@@ -99,7 +154,10 @@ if ModOptions and ModOptions.getInstance then
     ModOptions:loadFile()
 
 else
-    CHC_settings.config.list_font_size = getText("UI_optionscreen_Large")
+    CHC_settings.config.recipe_selector_modifier = 1
+    CHC_settings.config.category_selector_modifier = 1
+    CHC_settings.config.tab_selector_modifier = 1
+    CHC_settings.config.list_font_size = 3
     CHC_settings.config.allow_special_search = true
     CHC_settings.config.show_icons = false
     CHC_settings.config.show_hidden = true
@@ -129,25 +187,22 @@ CHC_settings.Load = function()
     if json and json ~= "" then
         CHC_settings.config = Json.Decode(json)
     else
-        local init_cfg = {
-            list_font_size = getText("UI_optionscreen_Large"),
-            show_icons = false,
-            allow_special_search = true,
-            show_hidden = true,
-            close_all_on_exit = false,
-            main_window = { x = 100, y = 100, w = 1000, h = 600 },
-            uses = { sep_x = 500, filter_asc = true, filter_type = "all" },
-            craft = { sep_x = 500, filter_asc = true, filter_type = "all" },
-            search = {
-                items = { sep_x = 500, filter_asc = true, filter_type = "all" },
-                recipes = { sep_x = 500, filter_asc = true, filter_type = "all" }
-            },
-            favorites = {
-                items = { sep_x = 500, filter_asc = true, filter_type = "all" },
-                recipes = { sep_x = 500, filter_asc = true, filter_type = "all" }
-            }
-        }
         CHC_settings.config = init_cfg
+        CHC_settings.Save()
+    end
+    CHC_settings.checkConfig()
+end
+
+
+CHC_settings.checkConfig = function()
+    local shouldReSave = false
+    for name, _ in pairs(init_cfg) do
+        if not CHC_settings.config[name] then
+            CHC_settings.config[name] = init_cfg[name]
+            shouldReSave = true
+        end
+    end
+    if shouldReSave == true then
         CHC_settings.Save()
     end
 end